Rename the project (blocklist.pf -> blockem)
