+++ include/blockem/cmd.h
