+++ include/blockem/file.h
@@ -1,6 +1,6 @@
 #pragma once
-#include <blocklist.pf/dyn_array.h>
-#include <blocklist.pf/set.h>
+#include <blockem/dyn_array.h>
+#include <blockem/set.h>
 dyn_array* read_file(const char *path);
 dyn_array* filter_file(dyn_array *file, struct Set *set);
 char* format_file(dyn_array *file, int per_line);
