+++ include/blockem/string.h
