+++ src/fetch.c
@@ -1,4 +1,4 @@
-#include <blocklist.pf/fetch.h>
+#include <blockem/fetch.h>
 
 dyn_array *
 fetch_blocklist(blocklist *bl)
