+++ src/file.c
@@ -1,7 +1,7 @@
-#include <blocklist.pf/dyn_array.h>
-#include <blocklist.pf/string.h>
-#include <blocklist.pf/set.h>
-#include <blocklist.pf/malloc.h>
+#include <blockem/dyn_array.h>
+#include <blockem/string.h>
+#include <blockem/set.h>
+#include <blockem/malloc.h>
 #include <unistd.h>
 #include <stdlib.h>
 #include <stdio.h>
