+++ src/malloc.c
@@ -1,4 +1,4 @@
-#include <blocklist.pf/malloc.h>
+#include <blockem/malloc.h>
 
 void *
 safe_malloc(size_t size)
